+++ packages/web3-providers-ws/CHANGELOG.md
@@ -122,11 +122,16 @@ Documentation:
 
 -   Dependencies updated
 
-
 ## [4.0.7]
 
 ### Fixed
 
 -   Fixed bug in chunks processing logic (#6496)
+
+## [4.0.8]
+
+### Changed
+
+-   Update dependancies (#7109)
 
 ## [Unreleased]